Language select and refactor
