+++ ModAssistant/MainWindow.xaml.cs
@@ -65,6 +65,9 @@ namespace ModAssistant
             Themes.LoadThemes();
             Themes.FirstLoad(Properties.Settings.Default.SelectedTheme);
 
+            // Load languages into combobox
+            Languages.LoadLanguages();
+
             Task.Run(() => LoadVersionsAsync());
 
             if (!Properties.Settings.Default.Agreed || string.IsNullOrEmpty(Properties.Settings.Default.LastTab))
