+++ ModAssistant/Pages/Options.xaml.cs
@@ -337,12 +337,14 @@ namespace ModAssistant.Pages
             {
                 // Apply default language
                 Console.WriteLine("Applying default language");
+                Languages.LoadLanguage("en");
             }
             else
             {
-                // Set new language
-                Console.WriteLine("Setting language to:" + (sender as ComboBox).SelectedItem.ToString());
-                //LoadLanguage((sender as ComboBox).SelectedItem.ToString())
+                // Get the matching language from the LoadedLanguages array, then try and use it
+                var languageName = (sender as ComboBox).SelectedItem.ToString();
+                var selectedLanguage = Languages.LoadedLanguages.Find(language => language.EnglishName.CompareTo(languageName) == 0);
+                Languages.LoadLanguage(selectedLanguage.Name);
             }
         }
